% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R
\name{build_kdtree_R}
\alias{build_kdtree_R}
\title{Construit un arbre KD}
\usage{
build_kdtree_R(points, depth = 0)
}
\arguments{
\item{points}{Un data.frame contenant les points et leurs classes. La dernière colonne est supposée être la classe.}

\item{depth}{La profondeur actuelle dans l'arbre, utilisée pour alterner les dimensions de split.}
}
\value{
Un objet `Node_R` représentant la racine de l'arbre KD construit.
}
\description{
Prend un ensemble de points et construit récursivement un arbre KD.
}
\examples{
data <- data.frame(x = runif(10), y = runif(10), class = sample(1:2, 10, replace = TRUE))
kd_tree <- build_kdtree_R(data)
}
