% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R, R/KNN_R.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calcule la Distance Euclidienne}
\usage{
euclidean_distance(row1, row2)

euclidean_distance(row1, row2)
}
\arguments{
\item{row1}{Un vecteur numérique représentant le premier point.}

\item{row2}{Un vecteur numérique représentant le deuxième point.}

\item{point1}{Un vecteur numérique représentant le premier point.}

\item{point2}{Un vecteur numérique représentant le second point.}
}
\value{
La distance euclidienne entre `point1` et `point2`.

Retourne un unique nombre réel représentant la distance euclidienne entre les deux points.
}
\description{
Calcule la distance euclidienne entre deux points.

Cette fonction calcule la distance euclidienne entre deux points, représentés par des vecteurs numériques.
}
\examples{
euclidean_distance(c(1, 2), c(3, 4))
euclidean_distance(c(1, 2), c(4, 6))
}
