% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_R.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Trouver les k Plus Proches Voisins}
\usage{
get_neighbors(training_set, test_instance, k)
}
\arguments{
\item{training_set}{Un data frame ou une matrice où chaque ligne est une instance d'entraînement
et les colonnes représentent les caractéristiques (la dernière colonne peut représenter la classe).}

\item{test_instance}{Un vecteur numérique représentant l'instance de test.}

\item{k}{Le nombre de voisins les plus proches à identifier.}
}
\value{
Un sous-ensemble de `training_set` contenant les k voisins les plus proches de l'instance de test.
}
\description{
Identifie les k plus proches voisins d'une instance de test dans un ensemble d'entraînement donné,
en utilisant la distance euclidienne comme mesure de proximité.
}
\examples{
data(iris)
test_instance <- iris[100, -5]
neighbors <- get_neighbors(iris, test_instance, k = 5)
}
