% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R
\name{nearest_neighbors_search_R}
\alias{nearest_neighbors_search_R}
\title{Recherche des Plus Proches Voisins dans un Arbre KD}
\usage{
nearest_neighbors_search_R(node, target_point, k, depth = 0, neighbors = NULL)
}
\arguments{
\item{node}{Le nœud racine de l'arbre KD.}

\item{target_point}{Le point cible pour lequel trouver les plus proches voisins.}

\item{k}{Le nombre de voisins à trouver.}

\item{depth}{La profondeur actuelle, utilisée pour choisir la dimension du split.}

\item{neighbors}{Une liste initialement vide pour stocker les voisins trouvés.}
}
\value{
Une liste contenant les k plus proches voisins du `target_point`.
}
\description{
Pour un point donné, trouve les k plus proches voisins dans un arbre KD.
}
\examples{
# Voir l'exemple de `build_kdtree_R` pour construire un arbre KD avant d'utiliser cette fonction.
}
