% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\alias{Node_R}
\title{Classe Node pour l'Arbre KD}
\description{
Cette classe représente un nœud dans un arbre KD, avec des champs pour le point (coordonnées et classe),
les fils gauche et droit, et la dimension utilisée pour le split.
}
\section{Fields}{

\describe{
\item{\code{point}}{Un vecteur numérique incluant les coordonnées et la classe du point.}

\item{\code{left}}{Le fils gauche du nœud, représentant l'espace de partitionnement à gauche ou en dessous du point de split.}

\item{\code{right}}{Le fils droit du nœud, représentant l'espace de partitionnement à droite ou au-dessus du point de split.}

\item{\code{dimension}}{La dimension (indice de colonne) utilisée pour le split à ce nœud.}
}}


\examples{
node <- Node_R$new(point = c(1.5, 2.5, 1), left = NULL, right = NULL, dimension = 1)
}
