% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R
\name{predict_class_KD_R}
\alias{predict_class_KD_R}
\title{Prédit la Classe d'un Point Utilisant l'Arbre KD}
\usage{
predict_class_KD_R(classes)
}
\arguments{
\item{classes}{Les classes des k plus proches voisins trouvés.}
}
\value{
La classe prédite pour le point cible.
}
\description{
Utilise un arbre KD pour prédire la classe d'un ensemble de points cibles en trouvant leurs plus proches voisins.
}
\examples{
# Voir l'exemple de `nearest_neighbors_search_R` pour obtenir les classes avant d'utiliser cette fonction.
}
