% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_R.R
\name{predict_classification_KNN}
\alias{predict_classification_KNN}
\title{Prédiction de Classe avec K Plus Proches Voisins}
\usage{
predict_classification_KNN(training_set, test_instance, k)
}
\arguments{
\item{training_set}{Un data frame ou une matrice de l'ensemble d'entraînement avec les dernières colonnes
représentant la classe de chaque instance.}

\item{test_instance}{Un vecteur numérique représentant l'instance de test.}

\item{k}{Le nombre de voisins à considérer pour la prédiction.}
}
\value{
La classe prédite pour l'instance de test, en tant qu'entier.
}
\description{
Prédit la classe d'une instance de test basée sur les votes majoritaires des k plus proches voisins
dans l'ensemble d'entraînement.
}
\examples{
data(iris)
test_instance <- iris[150, -5]
class_predicted <- predict_classification_KNN(iris, test_instance, k = 5)
}
