% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_kdtree_cpp.R
\name{predict_kdtree_cpp}
\alias{predict_kdtree_cpp}
\title{Prédit la classe pour des points cibles en utilisant un KD-tree}
\usage{
predict_kdtree_cpp(dataset, target_points, k)
}
\arguments{
\item{dataset}{Un data frame ou une matrice représentant l'ensemble de données d'entraînement.}

\item{target_points}{Un data frame ou une matrice des points cibles pour lesquels prédire les classes.}

\item{k}{Le nombre de voisins les plus proches à considérer pour la prédiction.}
}
\value{
Un vecteur numérique des classes prédites pour chaque point cible.
}
\description{
Cette fonction utilise un KD-tree implémenté en C++ pour prédire la classe
de points cibles donnés, basée sur un ensemble d'entraînement.
}
\examples{
# Exemple d'utilisation ici
}
