% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDtree.R
\name{predict_kdtree_R}
\alias{predict_kdtree_R}
\title{Prédit les Classes pour Plusieurs Points Utilisant l'Arbre KD}
\usage{
predict_kdtree_R(dataset, target_points, k)
}
\arguments{
\item{dataset}{Le dataset d'entraînement utilisé pour construire l'arbre KD.}

\item{target_points}{Un data.frame de points cibles.}

\item{k}{Le nombre de plus proches voisins à considérer pour la prédiction.}
}
\value{
Un vecteur des classes prédites pour chaque point cible.
}
\description{
Prédit les Classes pour Plusieurs Points Utilisant l'Arbre KD
}
\examples{
# Voir l'exemple de `build_kdtree_R` pour construire un arbre KD avant d'utiliser cette fonction.
}
