% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNRcpp.R
\name{predict_knn_cpp}
\alias{predict_knn_cpp}
\title{Predict class labels for target points using the KNN algorithm implemented in C++}
\usage{
predict_knn_cpp(trainingSet, target_points, k)
}
\arguments{
\item{trainingSet}{A matrix or data frame of training points where the last column is the class label.}

\item{target_points}{A matrix or data frame of target points for which to predict the classes.}

\item{k}{The number of nearest neighbors to consider.}
}
\value{
A numeric vector of predicted class labels for the target points.
}
\description{
This function takes a training set and a set of target points, prepares the data, and then
calls the C++ implemented KNN function for each target point to predict their classes.
}
\examples{
trainingSet <- matrix(c(1, 2, 1, 3, 4, 2, 5, 6, 1, 7, 8, 2), ncol=3, byrow=TRUE)
target_points <- matrix(c(2, 3, 4, 5), ncol=2, byrow=TRUE)
k <- 3
predictions <- predict_knn_cpp(trainingSet, target_points, k)
}
